[ui-components] refactor: added infoicon
