+++ packages/plugin-commons/web/src/Base/createBaseComponent.jsx
@@ -10,7 +10,7 @@ import {
   textWrapClassName,
   createHocName,
 } from 'wix-rich-content-common';
-import styles from 'wix-rich-content-editor-common/dist/statics/styles/general.scss';
+import styles from 'wix-rich-content-ui-components/dist/statics/styles/general.scss';
 import rtlIgnoredStyles from 'wix-rich-content-common/dist/statics/styles/general.rtlignore.scss';
 
 const DEFAULTS = Object.freeze({
