+++ packages/plugin-social-polls/web/src/components/settings/poll-settings-section/PollSettingsSection.jsx
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
-import { LabeledToggle } from 'wix-rich-content-ui-components';
-import { Separator, RadioGroup, InfoIcon } from 'wix-rich-content-editor-common';
+import { LabeledToggle, InfoIcon } from 'wix-rich-content-ui-components';
+import { Separator, RadioGroup } from 'wix-rich-content-editor-common';
 import { mergeStyles } from 'wix-rich-content-common';
 
 import { MEMBER_ROLES, VISIBILITY } from '../../../defaults';
