+++ packages/ui-components/web/src/Components/InfoIcon.jsx
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import generalstyles from '../../statics/styles/general.scss';
+import generalStyles from '../../statics/styles/general.scss';
 import Tooltip from 'wix-rich-content-common/libs/Tooltip';
 import Icon from '../Icons/InfoIcon';
 
 class InfoIcon extends Component {
   render() {
     const { tooltipText, iconStyles } = this.props;
-    const style = iconStyles || generalstyles.infoIcon;
+    const style = iconStyles || generalStyles.infoIcon;
     return (
       <Tooltip content={tooltipText}>
         <Icon className={style} />
