+++ packages/ui-components/web/statics/styles/general.scss
